+++ includes/optional-modules/class-rss.php
@@ -817,29 +817,37 @@ class RSS {
 			$custom_tracking_snippet,
 			[
 				'script'   => [
-					'id'    => true,
-					'src'   => true,
-					'type'  => true,
-					'async' => true,
-					'defer' => true,
-					'class' => true,
+					'id'          => true,
+					'src'         => true,
+					'type'        => true,
+					'async'       => true,
+					'defer'       => true,
+					'class'       => true,
+					'crossorigin' => true,
+					'data-*'      => true,
 				],
 				'img'      => [
-					'id'     => true,
-					'style'  => true,
-					'src'    => true,
-					'alt'    => true,
-					'class'  => true,
-					'width'  => true,
-					'height' => true,
+					'id'          => true,
+					'style'       => true,
+					'src'         => true,
+					'alt'         => true,
+					'class'       => true,
+					'width'       => true,
+					'height'      => true,
+					'data-*'      => true,
+					'crossorigin' => true,
+					'loading'     => true,
 				],
 				'iframe'   => [
-					'id'     => true,
-					'style'  => true,
-					'src'    => true,
-					'class'  => true,
-					'width'  => true,
-					'height' => true,
+					'id'          => true,
+					'style'       => true,
+					'src'         => true,
+					'class'       => true,
+					'width'       => true,
+					'height'      => true,
+					'crossorigin' => true,
+					'data-*'      => true,
+					'loading'     => true,
 				],
 				'noscript' => true,
 				'div'      => true,
@@ -1125,7 +1133,7 @@ class RSS {
 			$thumbnail_url = get_the_post_thumbnail_url( $post, 'full' );
 			if ( $thumbnail_url ) :
 				?>
-				<image><?php echo esc_url( $thumbnail_url, null, 'db' ); ?></image>
+				<image><?php echo esc_url( $thumbnail_url ); ?></image>
 				<?php
 			endif;
 		}
@@ -1155,11 +1163,11 @@ class RSS {
 				if ( $thumbnail_data ) {
 					$caption = get_the_post_thumbnail_caption();
 					?>
-					<media:content type="<?php echo esc_attr( get_post_mime_type( $thumbnail_id ) ); ?>" url="<?php echo esc_url( $thumbnail_data[0], null, 'db' ); ?>">
+					<media:content type="<?php echo esc_attr( get_post_mime_type( $thumbnail_id ) ); ?>" url="<?php echo esc_url( $thumbnail_data[0] ); ?>">
 						<?php if ( ! empty( $caption ) ) : ?>
 						<media:description><?php echo esc_html( $caption ); ?></media:description>
 						<?php endif; ?>
-						<media:thumbnail url="<?php echo esc_url( $thumbnail_data[0], null, 'db' ); ?>" width="<?php echo esc_attr( $thumbnail_data[1] ); ?>" height="<?php echo esc_attr( $thumbnail_data[2] ); ?>" />
+						<media:thumbnail url="<?php echo esc_url( $thumbnail_data[0] ); ?>" width="<?php echo esc_attr( $thumbnail_data[1] ); ?>" height="<?php echo esc_attr( $thumbnail_data[2] ); ?>" />
 					</media:content>
 					<?php
 				}